Move schoolSectionUtils from utils folder to forms folder Move helper functions in utils under helperFunctions folder
