+++ libs/application/templates/secondary-school/src/shared/enums.ts
@@ -1,0 +1,4 @@
+export enum ApplicationType {
+  FRESHMAN = 'FRESHMAN',
+  GENERAL_APPLICATION = 'GENERAL_APPLICATION',
+}
