+++ libs/application/templates/secondary-school/src/shared/index.ts
@@ -1,0 +1,1 @@
+export * from './enums'
