+++ libs/application/templates/secondary-school/src/utils/enums/index.ts
@@ -37,11 +37,6 @@ export enum Routes {
   SCHOOL = 'schoolMultiField',
 }
 
-export enum ApplicationType {
-  FRESHMAN = 'FRESHMAN',
-  GENERAL_APPLICATION = 'GENERAL_APPLICATION',
-}
-
 export enum ConclusionView {
   DEFAULT = 'DEFAULT',
   OVERVIEW = 'OVERVIEW',
