+++ libs/application/templates/secondary-school/src/utils/helperFunctions/checkHasAnyCustodians.ts
