+++ libs/application/templates/secondary-school/src/utils/helperFunctions/checkIsActor.ts
