+++ libs/application/templates/secondary-school/src/utils/helperFunctions/checkIsEditable.ts
@@ -1,4 +1,4 @@
-import { States } from './enums'
+import { States } from '../enums'
 
 export const checkIsEditable = (state: string) => {
   return state === States.DRAFT || state === States.EDIT
