+++ libs/application/templates/secondary-school/src/utils/helperFunctions/checkIsFreshman.ts
@@ -1,6 +1,6 @@
 import { getValueViaPath } from '@island.is/application/core'
-import { ApplicationType } from './enums'
 import { FormValue } from '@island.is/application/types'
+import { ApplicationType } from '../../shared'
 
 export const checkIsFreshman = (answers: FormValue): boolean => {
   return (
