+++ libs/application/templates/secondary-school/src/utils/helperFunctions/format.ts
