+++ libs/application/templates/secondary-school/src/utils/helperFunctions/hasDuplicates.ts
