+++ libs/application/templates/secondary-school/src/utils/helperFunctions/index.ts
@@ -1,0 +1,8 @@
+export * from './checkHasAnyCustodians'
+export * from './format'
+export * from './getDate'
+export * from './hasDuplicates'
+export * from './checkIsEditable'
+export * from './checkUseAnswersCopy'
+export * from './checkIsFreshman'
+export * from './checkIsActor'
