+++ libs/application/templates/secondary-school/src/utils/index.ts
@@ -1,12 +1,4 @@
 export * from './constants'
 export * from './enums'
 export * from './types'
-export * from './checkHasAnyCustodians'
-export * from './format'
-export * from './getDate'
-export * from './hasDuplicates'
-export * from './checkIsEditable'
-export * from './checkUseAnswersCopy'
-export * from './checkIsFreshman'
-export * from './checkIsActor'
-export * from './schoolSelection'
+export * from './helperFunctions'
